Cleaned some code and fixed a bug
Splat the Note classes to assets\Code\Class\NoteClasses.py
Fixed characters not correctly inverting their position when playing as opponent
